Bin: Implement State::register_event() and State::deregister_event()
These functions add a match rule on the interface member of the
specified event, and register our interest in this event with the atspi
registryd, in the correct order.

One current limitation is that duplicate matches will be added for
events with the same interface member. If this is a problem, we can
record which match rules we've already added, and only add new ones, and
remove existing ones.

Another limitation is that currently, these functions don't let you
specify a specific path, sender, or destination for the match rules.
This could easily be added in future if it's required.