Atspi: Create a new Event type
This will be yielded by an event stream in the future. It necessitated
renaming atspi::events::Event to atspi::events::EventBody, which is more
appropriate.

For now, this implementation uses the newly implemented EventBodyOwned
structure, and thus copies the event's kind and any_data. This could be
improved in the future, but it's more than good enough for now.