remove color codes from terminal output
The `tracing_tree` crate uses ancy color codes in the log output, disregarding the possibility that the resulting text may be piped to a file.
In the case of redirection to a non-tty output, such as through a pipe to a file or socket,
the color codes are shown as raw values in the text itself,
which is of course undesirable for various reasons, one of them is that the
color codes are interpreted as characters by most text editors, causing useless noise.
Untill we find a better solution, we disabled ancy color codes alltogether