Speech: Use speech_dispatcher crate
This is not fully optimal for our async-based architecture, but it's
perfectly usable for now. I chose the threaded mode so that any blocking
calls happen in the worker thread and don't block the executer, though
this may not actually be the case. Either way, if we *do* block, it will
only be until speechd reads some events.