Begin work on dispatching events
Currently we only dispatch Object:StateChanged:focused, and ignore the
rest. Ignored events are logged.

Currently events are processed serially. In future we plan to process
events in parallel, however, simply spawning a new task for each event
could lead to various issues and inefficiencies (I.E. lock contension,
processing events that are obsilete due to the receipt of newer events,
etc), so this will need to be implemented with care.