common: add configuration primitives
With the help of config, this adds a skeleton up on which more advanced configuration options could be built.
For now, the skeleton has the following parts:

* a central settings module, which contains the `ApplicationSettings` type.
This type is the main hub of all configuration sections in odilia because it has a field for each section defined in `common/settings/*.rs`.
Furthermore, it represents a read-only view of the configuration file, therefore only getter functions are provided to retrieve an immutable reference to a section.
If modifying values inside the configuration is intended, the file should be changed, then the current instance of `ApplicationSettings` should be replaced by a new call to `ApplicationSettings::new(path)`.
* a mod file in `common/settings/` for each section of odilia we want to make configurable
    * `common/settings/speech.rs` contains the `SpeechSettings` type, which represents all the configurable settings we want to expose for controlling speech. Only `speech_rate` is present there at the moment
    * `common/settings/log.rs` contains the `LogSettings` type which controlls all aspects of logging. For now, only `level` is present

For now, the `config` crate is used to gather configuration values from the following sources:

* the file given in the `path` parameter. For now, it can be only in the toml, yaml or json format
* from the environment, any environment variable prefixed by odilia. For example, `ODILIA_LOG_LEVEL=info ./target/odilia` will set the level property of log to info

additional information:

the sections also derive `Serde::Serialize`, even though we might not need it for the job of simply reading from a configuration file.
The reason for this is that `common` is depended upon by multiple crates, possibly including the configuration writer, so it's convenient for the writer to share this model with the config reader, with the added benefit of the writer being able to write changed values back to the file.