bin: register interest for the Object:TextCaretMoved atspi event
when this event is raised, the toolkit sygnals the screenreader that the caret, or edit cursor as it's also called, is moving.
This can only happen when cursor moving keys are pressed, for example arrow keys.
Handling this event properly will allow odilia to read text as the user navigates through it, as well as newly typed characters.