bin: add event handler for Object:TextCaretMoved
this handler doesn't do much, it's just placeholder code.
However, it speaks the position of the carret as recieved from event.detail1().
Also, the number increases when typing new characters or moving to a new line and decreases when deleting characters or going a line upwards, so the event recieving part seemns to work as expected