Use rustc-hash for cache evmap
The rustc-hash crate provides a hasher that is faster than sipHash
(std's built-in hasher) as it's less DoS resistent (something we don't
need), and even faster than the FNV hash, at the cost of slightly more
hash collisions. It was already in our dependency graph, so given its
performance and use by Mozilla and the Rust compiler, I decided it was
the best option.