Bin: Spawn main futures as tasks
tokio::join!() runs the futures you give it concurrently on the same
task, meaning only one core is used. To fix this, we spawn the futures
that process at-spi events and update the screen reader mode on their
own Tokio tasks, so they're free to be assigned to different cores. The
third task (processing Odilia events), is kept on the main task to
reduce the overhead of spawning tasks.

This required taking ownership of the ends of a communication channel
and the stream of Odilia events, but this should've been the case
anyway.

Some magic has to be done in order to convert all the error types
returned by these functions (including tokio::spawn()) into
eyre::Result<()>. We do this so that tokio::try_join!() works properly
and will correctly abort early if any of the futures or tasks fail.