[atspi][clippy] replace `i32` to `usize` casts into `try_into`
Clippy suggests eliminating casts in the `get_siblings_*` functions.

The `i32` is produced by `Accessob;e{rpxy::get_index_in_parent` which, according to:
[AT-SPI2 API docs](https://docs.gtk.org/atspi2/method.Accessible.get_index_in_parent.html),
may return -1 when the object has no parent or in case of an exception.

This proposes a `try_from` approach, which has as downside is that it chagnes the relevant trait method signatures as I had to 'strip the type' from the Result.

Would it perhaps be better to introduce an `AtspiError` type that makes clear that this error is a 'NoParentOrException' error?