[atspi][clippy] documentation item is missing backticks
The generated documentation did not backtick `DBus`.

I Complied to Clippy's wishes even though DBus is not necessarily a code item, but a protocol name.
I am not quite sure it it should be backtaicked at all.

Alternatively, we could just revert this commit and `allow(doc_markdown)` somewhere appropriate.