Hotfix after initial release
this fixes a bug which happened because the code didn't update the variable which was supposed to track the previous position of the caret. In consequence, the bug manifests to the user as if that event handler was programmed to fetch all the data in the current text accessible, from the first character, up untill the current caret position, that's why whenever one would type in an edit field, one would hear odilia say everything typed before.