add clippi warning for print statements
From now on, any code that contains the print! and println! macros will not be able to be merged because CI would fail on clippy lints, to fix the issue, use debug! or info! in stead. This is very good because sometimes we accidentally put one of those in a function or something, since we're used to it and it entered in muscle memory. While a local lint would also be nice through something like a clippy.toml file so lsp errors on us while building the code, that's the best we got for now