Move complex cache types into odilia_cache
It turns out we need async locking primitives, which at this time can not be compiled for any WASM architechture.
This restriction *should* be able to be lifted in the future, especially with the rise of two features:

1. An asyncronous WASM standard.
2. Reference types in WASM.

With both these features implemented, it is reasonably likely that no (de)serialization would have to be performed at all, and plugins could simply use the types as if they were a native Rust function.
This functionality, or event standardization is likely to take over a year, so I'm not holding my breath out for it, but when it comes, we should move the more advanced caching types back into odilia_common, and allow plugins free access to the cache, and maybe event the entire state, completely by reference.