RwLock -> Mutex
Most operation in Odilia need to be done as an "atomic grouping" if you will.
For example, suppose two events come in: the first sets some new text, the second moves the cursor over the newly inserted word.
Although in reality it will be *incredibly* rare that these two events happen fast enough for Odilia to have this problem, in theory the section of text to speak could be processed before the write is complete from the first event.

Overall, I'm just trying to make sure that if Odilia is going to have synchronization errors, at least they will be contained to receiving the events in the wrong order, and not ones self-imposed.