Fix-should-panic-lint (#127)
this fixes all issues reported by the failing clippy CI job, which prevents the all checks passed status for pull requests, making it harder for maintainers to distinguish between errors produced by clippy which are to be ignored and serious problems introduced by new code. In doing this, the following actions have been taken:

* tts: remove expect from spawning speech dispatcher

in stead, the message has been wrapped in `.context()`, which wraps the originating error in an anyhow error with the specified message, so that, in stead of panicking, the error gets bubbled up, logged and tracked, and then the program exits with the span traces belonging to that error.
this is much cleanner and healthier than panicking, so that's the reason I'm going with this

* input: don't use expect when opening listener socket

in stead, wrap the message with `anyhow::context`, then bubble it upward

* events and tests: remove unwraps and expects

in stead, anyhow::context is being used, to wrap the error
the signature of a function was changed to return a result, the test did likewise have to change to accomodate it. Luckily, that function was used only in one test

* object event test: annotated unwrap with allow

this is done because I can't meaningfully use error propagation in that context, since the unwrapping happens inside the declaration of a lazy static, but the explicit allow had to be added in order to eliminate all clippy warnings

* fix some more clippy lints

mostly related to using try_into instead of into, add explicit matching on unit types in stead of _, etc