timeout all the things
this introduces a timeout of hardcoded 500 milliseconds to the future which awaits all tasks to complete.
If, for example, a task is stuck somewhere and doesn't answer to cancelation tokens, we will exit the program forcefully, with an error code and the error will be displayed before exitting.
I believe around 500 milliseconds is enough for all tasks to get unstuck, but this will eventually be changeable within the configuration file.
For now though, the odilia is hanging on exit problem should probably be relatively well fixed with this