fix: screenreader applies configuration properly
the method join replaced the previously used merge. Apparently, with merge, when conflicts are encountered, it preferes keeping the current value, instead of replacing it. So, config was taken from the default values supplied with the modules, but the values in there were never replaced by those who should have a greatter priority.
as a consequence, user defined configuration files wouldn't be applied, and the user would understandably be confused by the results