Implement generic first parameters for Handler
Combined with the new atspi_event_handler function, automatically
convert a HandlerService for a specific event type, into a generic event
handler, which checks if the event type matches before passing it into
the inner event handler.

It will either: transform the event into its specific type, or abort
the calling of the service with an
`OdiliaError::AtspiError(AtspiError::Conversion(...))`

Clipppy and fmt

Use traits!