add a way for odilia to speak the roles of accessible controls as well as controls with more complicate labeling
For the labeling, we first check if the text is empty. If so, we check both name and description. If that's empty too, we look through the relationship set of objects, to see if there any labeled by relationships, and if so, extract the text from those labels instead

Some useless commented code was also removed, as well as other  general cleanup