fix code warnings and clean up the code base (#156)
* rename some structs in the cache_event module so that the warning about item starting with a partial name of the parent module is eliminated

* fix the too many lines warning by explicitly allowing it for the macro generated fun ction

* fix first wave of warnings

* formatting

* remove unreferenced macro in tests module

since the  tests in object.rs were removed in previous commits because they depended on functions marked as unused, this macro remained entirely unused, so it got removed

Because most of the functions in those modules are marked unused with the recent refactor, the tests testing them are equally invalid now, so better to make new, more relevant tests later, upon existing code

* fix wave 2 of warnings

* fix final wave of warnings

* remove the object module entirely, as none of its functions are needed

* remove the modes and screenreader events machinery, as it's not being used anyway and probably a better way with tower can be found

* formatting fixes, blame my code editor for modifying files in which I wasn't directly working

* add a way for odilia to speak the roles of accessible controls as well as controls with more complicate labeling

For the labeling, we first check if the text is empty. If so, we check both name and description. If that's empty too, we look through the relationship set of objects, to see if there any labeled by relationships, and if so, extract the text from those labels instead

Some useless commented code was also removed, as well as other  general cleanup